Move a few indexing utils to IndexingUtil
